+++ tests/testthat/test-test_indirect.R
@@ -47,59 +47,8 @@ sim_only <- power4test(nrep = 5,
                        do_the_test = FALSE,
                        iseed = 1234)
 
-#' @noRd
-
-test_indirect <- function(fit = fit,
-                          x = NULL,
-                          m = NULL,
-                          y = NULL,
-                          mc_ci = TRUE,
-                          mc_out = NULL,
-                          ...,
-                          get_map_names = FALSE,
-                          get_test_name = FALSE) {
-  map_names <- c(fit = "fit",
-                mc_out = "mc_out")
-  if (get_map_names) {
-    return(map_names)
-  }
-  if (get_test_name) {
-    tmp <- paste0(c(x, m, y),
-                  collapse = "->")
-    args <- as.list(match.call())
-    tmp2 <- character(0)
-    if (isTRUE(args$standardized_x) && !isTRUE(args$standardized_y)) {
-      tmp <- paste0(tmp, " ('x' standardized)")
-    }
-    if (!isTRUE(args$standardized_x) && isTRUE(args$standardized_y)) {
-      tmp <- paste0(tmp, " ('y' standardized)")
-    }
-    if (isTRUE(args$standardized_x) && isTRUE(args$standardized_y)) {
-      tmp <- paste0(tmp, " ('x' and 'y' standardized)")
-    }
-    return(paste0("test_indirect: ", tmp, collapse = ""))
-  }
-  out <- manymome::indirect_effect(x = x,
-                                   y = y,
-                                   m = m,
-                                   fit = fit,
-                                   mc_ci = mc_ci,
-                                   mc_out = mc_out,
-                                   progress = FALSE,
-                                   ...)
-  ci0 <- stats::confint(out)
-  out1 <- ifelse((ci0[1, 1] > 0) || (ci0[1, 2] < 0),
-                  yes = 1,
-                  no = 0)
-  out2 <- c(est = unname(coef(out)),
-            cilo = ci0[1, 1],
-            cihi = ci0[1, 2],
-            sig = out1)
-  return(out2)
-}
-
 test_ind <- power4test(object = sim_only,
-                       test_fun = test_indirect,
+                       test_fun = test_indirect_effect,
                        test_args = list(x = "x",
                                         m = "m",
                                         y = "y",
@@ -109,7 +58,7 @@ test_ind <- power4test(object = sim_only,
 expect_true(length(chk) == 1)
 
 test_ind2 <- power4test(object = test_ind,
-                       test_fun = test_indirect,
+                       test_fun = test_indirect_effect,
                        test_args = list(x = "x",
                                         y = "m",
                                         mc_ci = TRUE))
@@ -118,7 +67,7 @@ test_ind2 <- power4test(object = test_ind,
 expect_true(length(chk) == 2)
 
 test_ind3 <- power4test(object = test_ind2,
-                       test_fun = test_indirect,
+                       test_fun = test_indirect_effect,
                        test_args = list(x = "m",
                                         y = "y",
                                         mc_ci = TRUE))
@@ -127,7 +76,7 @@ test_ind3 <- power4test(object = test_ind2,
 expect_true(length(chk) == 3)
 
 test_ind4 <- suppressWarnings(power4test(object = test_ind3,
-                                         test_fun = test_indirect,
+                                         test_fun = test_indirect_effect,
                                          test_args = list(x = "x",
                                                            m = "m",
                                                            y = "y",